backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1
psycopg2-binary==2.9.9
watchdog==4.0.0
gunicorn==21.2.0
